+++ lib/scraped_page/strategy/live_request.rb
@@ -1,0 +1,23 @@
+require 'open-uri'
+
+class ScrapedPage
+  module Strategy
+    class LiveRequest
+      def response(url)
+        log "Fetching #{url}"
+        response = open(url)
+        {
+          status:  response.status.first.to_i,
+          headers: response.meta,
+          body:    response.read,
+        }
+      end
+
+      private
+
+      def log(message)
+        warn "[#{self.class}] #{message}"
+      end
+    end
+  end
+end
